requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
coveralls==1.8.1
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-nose==1.4.6
djangorestframework==3.9.1
djangorestframework-camel-case==1.0.3
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
logilab-astng==0.24.3
logilab-common==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
nose==1.3.7
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
mysqlclient==1.4.6